requirements.txt
numpy
pandas
torch
transformers
sentence-transformers
xgboost
scikit-learn
joblib
PyYAML
tqdm
jinja2
